requirements.txt
asgiref==3.8.1
bleach==6.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
django-bootstrap5==24.2
django-countries==7.6.1
django-markdownfield==0.11.0
django-phonenumber-field==7.3.0
django-pwa==2.0.1
gunicorn==22.0.0
idna==3.7
Markdown==3.7
mysqlclient==2.2.5
packaging==24.1
phonenumbers==8.13.38
pillow==10.3.0
python-decouple==3.8
pyuca==1.2
requests==2.32.3
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
tinycss2==1.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.6.0
